+++ controller/src/main/java/com/linbit/linstor/api/rest/v1/serializer/Json.java
@@ -106,6 +106,7 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
+import java.util.Objects;
 import java.util.TreeMap;
 import java.util.stream.Collectors;
 
@@ -574,6 +575,7 @@ public class Json
                                 vlmState = new JsonGenTypes.VolumeState();
                                 SatelliteVolumeState satVlmState = satResState.getVolumeStates().get(vlmNumber);
                                 vlmState.disk_state = satVlmState.getDiskState();
+                                vlmState.replication_state = Objects.toString(satVlmState.getReplicationState(), null);
                             }
                         }
                     }