+++ server/src/main/java/com/linbit/linstor/InternalApiConsts.java
@@ -100,6 +100,7 @@ public class InternalApiConsts
      * Events
      */
     public static final String EVENT_VOLUME_DISK_STATE = "VlmDiskState";
+    public static final String EVENT_REPLICATION_STATE = "ReplicationState";
     // State of resource based on DRBD status (or other underlying system)
     public static final String EVENT_RESOURCE_STATE = "ResourceState";
     public static final String EVENT_CONNECTION_STATE = "ConnectionState";