+++ server/src/main/java/com/linbit/linstor/satellitestate/SatelliteVolumeState.java
@@ -1,8 +1,11 @@
 package com.linbit.linstor.satellitestate;
 
+import com.linbit.linstor.layer.drbd.drbdstate.ReplState;
+
 public class SatelliteVolumeState
 {
     private String diskState;
+    private ReplState replicationState;
 
     public SatelliteVolumeState()
     {
@@ -11,6 +14,7 @@ public class SatelliteVolumeState
     public SatelliteVolumeState(SatelliteVolumeState other)
     {
         diskState = other.diskState;
+        replicationState = other.replicationState;
     }
 
     public String getDiskState()
@@ -23,8 +27,18 @@ public class SatelliteVolumeState
         diskState = diskStateRef;
     }
 
+    public ReplState getReplicationState()
+    {
+        return replicationState;
+    }
+
+    public void setReplicationState(ReplState replicationStateRef)
+    {
+        this.replicationState = replicationStateRef;
+    }
+
     public boolean isEmpty()
     {
-        return diskState == null;
+        return diskState == null && replicationState == null;
     }
 }